Module xpanda_cli::read

Functions

read_input_file

read_line

Reads the next line from stdin just like std::io::Lines::next except that it includes the line ending in the returned string.

read_named_arg

Tries to read a string in key=value format, returning the key and value as a tuple (in that order).

read_output_file

read_var_file

Reads a file of key=value pairs, ignoring empty lines.